Add ChezyInterruptHandlerFunction
This is a hack to fix a (once again) broken piece of WPILib.

WPILib has a class called InterruptHandlerFunction that is
supposed to let you specify a generic callback to happen
on an interrupt. The implementation includes no scope
modifier on the abstract method declaration and is therefore
only able to be overriden inside its own package.

This is a shim that lives in the wpilibj package that will
allow us to build interrupt handlers in other places in the code.